add favorite toggle functionality
