+++ src/hooks/useCubes.ts
@@ -1,12 +1,13 @@
 import { useEffect, useState } from "react";
 import { useTimerStore } from "@/store/timerStore";
-import { getAllCubes } from "@/db/dbOperations";
+import { getAllCubes, saveCube } from "@/db/dbOperations";
 import { Cube } from "@/interfaces/Cube";
+import { useRouter } from "@/i18n/routing";
 
 export function useCubes() {
-  const { cubes } = useTimerStore();
+  const { cubes, setSelectedCube, setNewScramble } = useTimerStore();
   const [filterCubes, setFilterCubes] = useState(cubes);
-
+  const router = useRouter();
   const handleSearchFilter = (searchCube: string) => {
     if (!cubes) return;
     if (searchCube.trim() === "") return setFilterCubes(cubes);
@@ -17,6 +18,34 @@ export function useCubes() {
     );
   };
 
+  const handleFavoriteClick = async (cubeId: string) => {
+    try {
+      const editingCube = cubes?.find((i) => i.id === cubeId);
+
+      if (!editingCube) return;
+
+      await saveCube({
+        ...editingCube,
+        favorite: !editingCube.favorite,
+      });
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  const redirectToTimer = (cubeId: string) => {
+    const clickedCube = cubes?.find((i) => i.id === cubeId);
+
+    // If the cube does not exist, do nothing
+    if (!clickedCube) {
+      return;
+    }
+
+    setSelectedCube({ ...clickedCube });
+    setNewScramble(clickedCube);
+    return router.push("/");
+  };
+
   useEffect(() => {
     getAllCubes().then((res) => setFilterCubes(res));
   }, [cubes]);
@@ -24,5 +53,7 @@ export function useCubes() {
   return {
     filterCubes,
     handleSearchFilter,
+    handleFavoriteClick,
+    redirectToTimer,
   };
 }
